save data in firebase

+++ app/src/main/java/com/example/bookverse/Class/Book.java
@@ -10,14 +10,14 @@ import java.util.ArrayList;
 import java.util.Map;
 
 public class Book implements Serializable {
-    private float id, download_count;
+    private int id, download_count;
     private String title, media_type;
     private String[] subjects, bookshelves, languages;
     private ArrayList<Person> authors, translators;
     private boolean copyright;
     private Map<String, String> formats;
 
-    public Book(ArrayList<Person> authors, String[] bookshelves, boolean copyright, float download_count, Map<String, String> formats, float id, String[] languages, String media_type, String[] subjects, String title, ArrayList<Person> translators) {
+    public Book(ArrayList<Person> authors, String[] bookshelves, boolean copyright, int download_count, Map<String, String> formats, int id, String[] languages, String media_type, String[] subjects, String title, ArrayList<Person> translators) {
         this.authors = authors;
         this.bookshelves = bookshelves;
         this.copyright = copyright;
@@ -55,11 +55,11 @@ public class Book implements Serializable {
         this.copyright = copyright;
     }
 
-    public float getDownload_count() {
+    public int getDownload_count() {
         return download_count;
     }
 
-    public void setDownload_count(float download_count) {
+    public void setDownload_count(int download_count) {
         this.download_count = download_count;
     }
 
@@ -71,11 +71,11 @@ public class Book implements Serializable {
         this.formats = formats;
     }
 
-    public float getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(float id) {
+    public void setId(int id) {
         this.id = id;
     }
